aris_resolver/requirements.txt
requests>=2.31.0
striprtf>=0.0.26
rapidfuzz>=3.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
python-docx>=1.0.0
